chore: move actions provider to separate file
